Support nested `.run` drivers
The `NVIDIA-Linux-x86_64-545.23.08.run` driver is currently exclusively
available with CUDA 12.3.1.

It can only be downloaded separately as a `.deb`/`.rpm` from
https://developer.download.nvidia.com/compute/cuda/repos.

If we want a `.run` file, it can only be located within the
`./builds/NVIDIA-Linux-x86_64-545.23.08.run` folder of the
`cuda_12.3.1_545.23.08_linux.run` file.

Luckily, it's pretty easy to handle, we just need to repeat the
`main()` function again if we spot a
`./builds/NVIDIA-Linux-x86_64-545.23.08.run` file, although it does
mean we need to download the entire CUDA package just for a driver.