Package the i386 driver on x86_64
This allows us to stop depending on the ancient 1.6 i386 SDK to package
the i386 driver.

With these changes, the i386 driver is now packaged on a x86_64 host.

Unfortunately, we now also have to include zlib and liblzma, because
the 24.08 SDK doesn't include static libraries (.a files) for them, and
that will likely increase the overall build time.