Revert "Extract just library files when on GL32"
By not installing the Vulkan and EGL ICDs in the GL32 extension, we ended
up breaking 32-bit Vulkan support for most applications that need it,
e.g. game launchers that run 32-bit Windows games via Wine + DXVK.

The breakage was not noticed by me when I introduced this change, because
my tests were only done using Flatpak'd Steam, which unluckily for me,
includes a workaround that prevents the breakage.

The workaround is simple and consists of adding the path of the Vulkan
ICD to the XDG_CONFIG_DIRS environment variable, thus allowing the
Vulkan Loader (libvulkan.so) to find it. [1]

Another breakage affected applications using the EGL API (e.g. Wine with
its native Wayland driver), which can also be worked around by modifying
the __EGL_VENDOR_LIBRARY_DIRS environment variable to make the EGL ICD
discoverable. [2]

For those reasons, I'm reintroducing the ICDs (and a few other files) to
the GL32 extension.

Perhaps in the future we can investigate a different approach, e.g. by
somehow integrating these workarounds directly in the Freedesktop runtime
instead, which would eliminate the need to patch each application
individually.

This reverts commit d55b7cb0e5269b5bd4b0d86ac8df2f77966451ce.

[1] https://vulkan.lunarg.com/doc/view/1.3.250.1/windows/LoaderDriverInterface.html#driver-discovery-on-linux
[2] https://github.com/NVIDIA/libglvnd/blob/606f6627cf481ee6dcb32387edc010c502cdf38b/src/EGL/icd_enumeration.md#icd-discovery