CI: Reduce build batches from 30 to 15
We keep running out of disk space while exporting the drivers. [1]

I first tried to work around this in c19af31e ("CI: Free up more disk
space"), but that wasn't enough, so let's try to reduce the number of\
build batches from 30 to 15 and see how it goes.

My guess is that this is happening now because the forked version of
flatpak-builder we use to build the drivers was recently modified [2]
to always export a .Sources package(?) for x86_64 builds, which seems
to require about twice as much disk space as before...

[1] https://github.com/flathub/org.freedesktop.Platform.GL.nvidia/actions/runs/17412754189/job/49435186818#step:6:5294
[2] https://github.com/flathub-infra/flatpak-builder/commit/9d54e13a6c042558ff2265f279d60e358f86246b